test(http): Move `testRenderExceptionSetsDisplayErrorsInDebugMode` from `ApplicationTest` to `ApplicationErrorHandlerTest` for better organization.
