+++ tests/http/stateless/ApplicationErrorHandlerTest.php
@@ -348,6 +348,68 @@ final class ApplicationErrorHandlerTest extends TestCase
 
     /**
      * @throws InvalidConfigException if the configuration is invalid or incomplete.
+     */
+    #[RequiresPhpExtension('runkit7')]
+    public function testRenderExceptionSetsDisplayErrorsInDebugMode(): void
+    {
+        @\runkit_constant_redefine('YII_ENV_TEST', false);
+
+        $initialBufferLevel = ob_get_level();
+
+        $_SERVER = [
+            'REQUEST_METHOD' => 'GET',
+            'REQUEST_URI' => 'site/trigger-exception',
+        ];
+
+        $originalDisplayErrors = ini_get('display_errors');
+
+        try {
+            $app = $this->statelessApplication(
+                [
+                    'components' => [
+                        'errorHandler' => [
+                            'discardExistingOutput' => true,
+                            'errorAction' => null,
+                        ],
+                    ],
+                ],
+            );
+
+            $response = $app->handle(FactoryHelper::createServerRequestCreator()->createFromGlobals());
+
+            self::assertSame(
+                500,
+                $response->getStatusCode(),
+                "Expected HTTP '500' for route 'site/trigger-exception'.",
+            );
+            self::assertSame(
+                'text/html; charset=UTF-8',
+                $response->getHeaderLine('Content-Type'),
+                "Expected Content-Type 'text/html; charset=UTF-8' for route 'site/trigger-exception'.",
+            );
+            self::assertSame(
+                '1',
+                ini_get('display_errors'),
+                "'display_errors' should be set to '1' when YII_DEBUG mode is enabled and rendering exception view.",
+            );
+            self::assertStringContainsString(
+                'yii\base\Exception: Exception error message.',
+                $response->getBody()->getContents(),
+                'Response should contain exception details when YII_DEBUG mode is enabled.',
+            );
+        } finally {
+            ini_set('display_errors', $originalDisplayErrors);
+
+            while (ob_get_level() < $initialBufferLevel) {
+                ob_start();
+            }
+
+            @\runkit_constant_redefine('YII_ENV_TEST', true);
+        }
+    }
+
+    /**
+     * @throws InvalidConfigException if the configuration is invalid or incomplete.
      *
      * @phpstan-param string[] $expectedContent
      */
@@ -606,4 +668,5 @@ final class ApplicationErrorHandlerTest extends TestCase
 
         $app->request->resolve();
     }
+
 }
